Pull out non-cache code from StaticPages module
The `StaticPages` module seems to mostly deal with caching except for when its
pulled into the `PagesController`, where it has some handling code for
rendering 404s if a template is not found.

As the `PagesController` is the only place this is used I think it mkaes sense
to extract it to that controller and re-purpose the `StaticPages` module as a
`StaticPageCache` module.
