+++ spec/controllers/concerns/static_page_cache_spec.rb
@@ -1,11 +1,11 @@
 require "rails_helper"
 
-describe StaticPages do
+describe StaticPageCache do
   subject { tester.tap(&:render) }
 
   let(:testing_class) do
     Class.new do
-      include StaticPages
+      include StaticPageCache
       attr_reader :cacheable_static_page
 
       def render
@@ -70,58 +70,6 @@ describe StaticPages do
 
     it do
       is_expected.to have_received(:expires_in).with expires_in, public: true
-    end
-  end
-
-  describe "#filtered_page_template" do
-    subject { tester.send :filtered_page_template }
-
-    let(:params) { { page: template } }
-
-    before { allow(tester).to receive(:params).and_return params }
-
-    context "with valid page template" do
-      let(:template) { "hello" }
-
-      it { is_expected.to eql "hello" }
-    end
-
-    context "with nested template" do
-      let(:template) { "hello/world" }
-
-      it { is_expected.to eql "hello/world" }
-    end
-
-    context "with invalid page template" do
-      let(:template) { "invalid!" }
-
-      it { expect { subject }.to raise_exception StaticPages::InvalidTemplateName }
-    end
-
-    context "with param linking to parent page" do
-      let(:template) { "../../secrets.txt" }
-
-      it { expect { subject }.to raise_exception StaticPages::InvalidTemplateName }
-    end
-
-    context "with file extension" do
-      let(:template) { "stories/how-i-got-into-teaching.html" }
-
-      it { is_expected.to eql "stories/how-i-got-into-teaching.html" }
-    end
-
-    context "with numbers in name" do
-      let(:template) { "stories/my-top-10" }
-
-      it { is_expected.to eql "stories/my-top-10" }
-    end
-
-    context "with custom page value" do
-      subject { tester.send :filtered_page_template, :story }
-
-      let(:params) { { story: "hello" } }
-
-      it { is_expected.to eql "hello" }
     end
   end
 end
